cgroup: extract cgroup rest api handler from daemon
Currently, the cgroup rest api handler is implemented as part of the daemon itself.

To further minimize the number of dependencies on the daemon, this commit
extracts the cgroup rest api handler into the cgroup manager hive cell in package
`pkg/cgroups/manager`.
